+++ src/libmongoc/src/mongoc/mongoc-crypto-openssl-private.h
@@ -31,11 +31,11 @@ BSON_BEGIN_DECLS
 int
 mongoc_crypto_openssl_pbkdf2_hmac_sha1 (mongoc_crypto_t *crypto,
                                         const char *password,
-                                        uint32_t password_len,
+                                        size_t password_len,
                                         const uint8_t *salt,
-                                        uint32_t salt_len,
+                                        size_t salt_len,
                                         uint32_t iterations,
-                                        uint32_t key_len,
+                                        size_t key_len,
                                         unsigned char *output);
 
 void
@@ -55,11 +55,11 @@ mongoc_crypto_openssl_sha1 (mongoc_crypto_t *crypto,
 int
 mongoc_crypto_openssl_pbkdf2_hmac_sha256 (mongoc_crypto_t *crypto,
                                           const char *password,
-                                          uint32_t password_len,
+                                          size_t password_len,
                                           const uint8_t *salt,
-                                          uint32_t salt_len,
+                                          size_t salt_len,
                                           uint32_t iterations,
-                                          uint32_t key_len,
+                                          size_t key_len,
                                           unsigned char *output);
 
 void
