+++ src/libmongoc/src/mongoc/mongoc-crypto-openssl.c
@@ -28,11 +28,11 @@
 int
 mongoc_crypto_openssl_pbkdf2_hmac_sha1 (mongoc_crypto_t *crypto,
                                         const char *password,
-                                        uint32_t password_len,
+                                        size_t password_len,
                                         const uint8_t *salt,
-                                        uint32_t salt_len,
+                                        size_t salt_len,
                                         uint32_t iterations,
-                                        uint32_t key_len,
+                                        size_t key_len,
                                         unsigned char *output)
 {
    return PKCS5_PBKDF2_HMAC (password, password_len, salt, salt_len, iterations, EVP_sha1 (), key_len, output);
@@ -97,11 +97,11 @@ cleanup:
 int
 mongoc_crypto_openssl_pbkdf2_hmac_sha256 (mongoc_crypto_t *crypto,
                                           const char *password,
-                                          uint32_t password_len,
+                                          size_t password_len,
                                           const uint8_t *salt,
-                                          uint32_t salt_len,
+                                          size_t salt_len,
                                           uint32_t iterations,
-                                          uint32_t key_len,
+                                          size_t key_len,
                                           unsigned char *output)
 {
    return PKCS5_PBKDF2_HMAC (password, password_len, salt, salt_len, iterations, EVP_sha256 (), key_len, output);
